breaking stuff up into more functions
