+++ nrm-artist/masks_functs.py
@@ -3,222 +3,95 @@ import matplotlib.pyplot as plt
 import astropy.io.fits as pyfits
 from design_class import *
 from scipy.optimize import curve_fit 
+from pupil_plane import *
+from redundancy_check import *
+from aperture_class import *
 
 ########################################################################################################
 ########################################################################################################
 ########################################################################################################
 
-def check_hole_cent(hcoords, aperture):
-    hy = hcoords[0]
-    hx = hcoords[1]
-    if aperture[hy, hx] == 0:
-        return False
-    else:
-        return True
-
-########################################################################################################
-########################################################################################################
-########################################################################################################
-
-def check_hole_overlap(hcoords, hcoords_list, hrad):
-    if len(hcoords_list) == 0:
-        return True
-    else:
-        for i in hcoords_list:
-            dist = np.sqrt((i[0] - hcoords[0])**2 + (i[1] - hcoords[1])**2)
-            if dist < 100 * hrad:
-                return False
-            else:
-                continue
-
-
-########################################################################################################
-########################################################################################################
-########################################################################################################
-
-def check_spiders_gaps(hcoords, hrad, aperture):
-    """ Check placement of hole
-
-    Called by add_hole(). Checks that a proposed hole doesn't overlap other holes or spiders or mirror segment edges, and that it falls within the Keck aperture. If a hole does not meet requirements, hole is discarded and add_hole() is called again. Repeats until an acceptable hole location is found.
-
-    Args:
-        coords (array): numpy array containing the proposed (x,y) coordinates. 
-        design (object): Mask design object.
-        rng (object): Random number generator.
-        aperture (array): numpy array containing a boolean mask of the Keck primary.
-    
-    Returns:
-        array: returns the accepted (x,y) coordinates
-    """
-
-    xvals = np.arange(0, 1090, 1).reshape([1, 1090])
-    yvals = np.flip(np.arange(0, 1090, 1).reshape([1090, 1]))
-    distances = np.sqrt((xvals - hcoords[1])**2 + (yvals - hcoords[0])**2)
-    distances[distances < 1.0] = 1000.
-    distances[distances < (100 * hrad)] = -100.0
-    if np.min(distances + 200*aperture) < -10.0:
-        return False
-    return True
-
-########################################################################################################
-########################################################################################################
-########################################################################################################
-
-def add_hole_cent(rng, coord_options):
+def pick_hole_cent(rng, coord_options):
     a = np.array(rng.integers(low=0, high=len(coord_options)))
-    hcoords = coord_options[a] # + [545, 545] # convert proposed hole center coords to coords in aperture array
     coord_options2 = np.delete(coord_options, a, 0)
-    return np.array(hcoords), coord_options2
+    return coord_options[a], coord_options2
 
 ########################################################################################################
 ########################################################################################################
 ########################################################################################################
 
-def add_hole(rng, aperture, my_design, hcoords_list):
-    """ Propose a new mask hole
+#def pick_hole(rng):
 
-    Generates a proposed hole (x,y) coordinate set, then calls check_placement() to check whether these coordinates are acceptable. If they are, then these coordinates are returned. 
-    
-    Args:
-        hrad (float): Hole radius in meters.
-        rng (object): Random number generator.
-        aperture (array): numpy array containing a boolean mask of the Keck primary.
-    
-    Returns:
-        array: Returns a numpy array of the accepted hole (x,y) coordinates.
-    """
-    i = 0
-    flag = 0
-    while i < 50:
-        coords = np.array(rng.integers(low=-545, high=545, size=2))
-        hcoords = coords + [545, 545] # convert proposed hole center coords to coords in aperture array
-        if check_hole_cent(hcoords, aperture) == False:
-            continue
-        if check_spiders_gaps(hcoords, my_design.hrad, aperture) == False:
-            continue
-        if check_hole_overlap(hcoords, hcoords_list, my_design.hrad) == False:
-            continue
-        #if len(hcoords_list) > 0:
-        #    tmp_xycoords = np.array(list(my_design.xy_coords_cm) + hcoords)
-        #    tmp_design = design(my_design.nholes, my_design.hrad)
-        #    tmp_design.xy_coords_cm = tmp_xycoords
-        #    tmp_design.get_xy_m() # convert (x,y) coords in cm to m
-        #    tmp_design.get_uvs()
-        #    if check_redundancy(tmp_design) == 1:
-        #        i += 1
-        #        continue
-        hcoords_list.append(hcoords)
-        #print('yay')
-        return np.array(hcoords), hcoords_list, flag
-    flag = 1
-    return np.array(hcoords), hcoords_list #, flag
 
 ########################################################################################################
 ########################################################################################################
 ########################################################################################################
 
-def check_redundancy(my_design, bw):
-    """ Check mask baselines for redundancy
-
-    Checks for any redundancy in the baselines of the proposed mask design. If redundancy is above 0%, reject the mask design.
-
-    Args:
-        my_design (object): An instance of design class. The proposed mask to be tested.
-        bw (float): Bandwidth of the instrument; default is to assume infinitely narrow band (optional)
-
-    Returns:
-        bool: Returns 1 if the mask has any redundancy, returns 0 if mask is fully non-redundant.
-    """
-    pmask = 25 # radius in pixels for 0.5m mask
-    #pmask = 31 # radius in pixels for 0.65m mask
-    #psc = 
-    lam = 5.23*1e-6 # wavelength to use in meters; central wavelength of the band
-    avg = 0
-    count = 0
-    fact = 2*np.pi*206265/lam
-    uv_rad = my_design.hrad # need to change this
-    n = 50000
-    ci = 0
-    for i in my_design.uv_coords:
-        cj = 0
-        b1 = i
-        for j in my_design.uv_coords:
-            b2 = j
-            if ci == cj:
-                cj += 1
-                continue
-            cj += 1
-            d1 = np.sqrt((b1[0] - b2[0])**2 + (b1[1] - b2[1])**2) # both are positive
-            d2 = np.sqrt((-b1[0] - b2[0])**2 + (-b1[1] - b2[1])**2) # one is negative
-            d = np.min([d1, d2])
-            if d < 2*uv_rad:
-                test_uvs = np.random.uniform(low=0, high=2*uv_rad, size=(2,n))
-                dist_b1 = np.sqrt((test_uvs[0,:] - uv_rad)**2 + (test_uvs[1,:] - uv_rad)**2)
-                dist_b2 = np.sqrt((test_uvs[0,:] - (uv_rad + d))**2 + (test_uvs[1,:] - (uv_rad))**2)
-                count1 = (dist_b1 <= uv_rad).sum()
-                count2 = 0
-                for q in range(n):
-                    if dist_b1[q] <= uv_rad and dist_b2[q] < uv_rad:
-                        count2 += 1
-                red = 100 * np.round(count2 / count1, 2)
-                avg += red
-            count += 1
-        ci += 1
-    if count == 0:
-        avg_f = 0
+def make_design(nholes, hrad, ap, bw=0, geometry='cent'):
+    rng = np.random.default_rng(seed=None) # set random number generator
+    n = 7
+    if nholes > n:
+        if geometry == 'cent':
+            timeout1 = 0
+            while 1:
+                if timeout1 > 1000:
+                    raise Exception(f'Timed out, could not find a non-redundant design.')
+                coord_options = ap.hcmb_coords * 1.0
+                my_design0 = design(n, hrad)
+                for i in range(n):
+                    hole, coord_options = pick_hole_cent(rng, coord_options)
+                    my_design0.add_hole(np.reshape(hole, [1, 2]), i)
+                my_design0.get_xy_m()
+                my_design0.get_uvs()
+                rcheck = check_redundancy(my_design0, bw)
+                if rcheck == 0:
+                    timeout2 = 0
+                    while 1:
+                        tn = x_choose_y(24-n, nholes-n)
+                        if timeout2 > tn:
+                            break
+                        coord_options2 = coord_options * 1.0
+                        my_design = design(nholes, hrad)
+                        for i in range(n):
+                            my_design.add_hole(my_design0.xy_coords_cm[i], i)
+                        for i in range(nholes-n):
+                            hole, coord_options2 = pick_hole_cent(rng, coord_options2)
+                            my_design.add_hole(hole, n+i)
+                        my_design.get_xy_m()
+                        my_design.get_uvs()
+                        rcheck2 = check_redundancy(my_design, bw)
+                        if rcheck2 == 0:
+                            return my_design
+                        timeout2 += 1
+                timeout1 += 1
+        #if geometry == 'random':
+            # stuff
+        else:
+            raise AttributeError(f'Invalid geometry.')
     else:
-        avg_f = avg / count
-    if avg_f > 0:
-        return 1
-    else:
-        return 0
+        timeout = 0
+        while 1:
+            my_design = design(nholes, hrad)
+            coord_options = ap.hcmb_coords * 1.0
+            if timeout > 1000:
+                raise Exception(f'Timed out, could not find valid design.')
+            if geometry == 'cent':
+                coord_options = ap.hcmb_coords * 1.0
+                for i in range(nholes):
+                    hole, coord_options = pick_hole_cent(rng, coord_options)
+                    my_design.add_hole(np.reshape(hole, [1, 2]), i)
+            #if geometry == 'random': # randomly place holes, but still avoid furniture
+                #hole = pick_hole(rng)
+            else:
+                raise AttributeError(f'Invalid geometry.')
+            my_design.get_xy_m() # convert (x,y) coords in cm to m
+            my_design.get_uvs() # calculate design uv coordinates
+            rcheck = check_redundancy(my_design, bw)  # check design for redundancy
+            if rcheck == 0: # if this statement is true, exit the loop and return our final design
+                return my_design
+            timeout += 1
 
-########################################################################################################
-########################################################################################################
-########################################################################################################
-
-#def replace_hole(hrad, rng, aperture, hcoords_list):
-    
-    
-########################################################################################################
-########################################################################################################
-########################################################################################################
-
-def plot_design(my_design, aperture):
-    """ Plots finished design
-
-    Generates a matplotlib plot of the Keck aperture with the finished mask design projected.
-
-    Args:
-        my_design (object): Instance of design class representing the finished mask design.
-        aperture (array): Numpy array generated from the Keck aperture file provided.
-    
-    """
-
-    hcoords = my_design.xy_coords_cm
-    aperture2 = aperture.file * 1.0
-
-    if np.ndarray.flatten(hcoords > 545).any():
-        print('warning: coords not centered! fixing...')
-        hcoords -= [545, 545]
-    
-    for i in range(1090):
-        for j in range(1090):
-            for a in range(my_design.nholes):
-                if np.sqrt(((i - 545) - hcoords[a, 0])**2 + ((j - 545) - hcoords[a, 1])**2) < (100 * my_design.hrad):
-                     aperture2[i, j] = 0
-    
-    plt.figure()
-    plt.imshow(aperture2)
-    plt.xticks([])
-    plt.yticks([])
-    plt.show()
-
-########################################################################################################
-########################################################################################################
-########################################################################################################
-
-def make_design(nholes, hrad, ap, return_vcoords=False, bw=0): 
+def make_design2(nholes, hrad, ap, return_vcoords=False, bw=0, geometry='cent'): 
     """ Generates mask design
 
     Generates a single non-redundant aperture mask design using the user-inputted number of holes and hole radius.
@@ -234,9 +107,6 @@ def make_design(nholes, hrad, ap, return_vcoords=False, bw=0):
     """
 
     while 1: # keep looping until we get a valid design
-        coord_options = ap.hcmb_coords * 1.0
-        my_design = design(nholes, hrad) # initialize design object
-        rng = np.random.default_rng(seed=None) # set random number generator
         for i in range(nholes): # keep adding and checking a single hole until it's acceptable
             my_design.xy_coords_cm[i, :], coord_options = add_hole_cent(rng, coord_options)
         my_design.get_xy_m() # convert (x,y) coords in cm to m
@@ -252,42 +122,6 @@ def make_design(nholes, hrad, ap, return_vcoords=False, bw=0):
             
 ########################################################################################################
 ########################################################################################################
-########################################################################################################   
-    
-def add_to_design(mask_design, diff, vcoords0):
-    #aperture = pyfits.getdata('/Users/kenzie/Desktop/CodeAstro/planet-guts/keck_aperture.fits') # set Keck primary aperture
-    vcoords = vcoords0
-    hcoords_list = mask_design.xy_coords_cm.tolist()
-    rng = np.random.default_rng(seed=None)
-    for i in range(diff):
-        temp_design = design(mask_design.nholes + 1, mask_design.hrad)
-        count = 0
-        while 1:
-            new_hole, hcoords_list, vcoords = add_hole_cent(rng, hcoords_list, vcoords)
-            temp_design.xy_coords_cm = np.array(hcoords_list)
-            temp_design.xy_coords_cm = np.append(temp_design.xy_coords_cm, np.reshape(new_hole, [1,2]), axis=0)
-            temp_design.get_xy_m() # convert (x,y) coords in cm to m
-            temp_design.get_uvs() # calculate design uv coordinates
-            rcheck = check_redundancy(temp_design)
-            if rcheck == 1: # if the new hole makes it redundant:
-                np.delete(temp_design.xy_coords_cm, temp_design.nholes-1, axis=0) # delete what I just added to temp
-                hcoords_list = mask_design.xy_coords_cm.tolist() # return hcoords_list to normal
-                vcoords = vcoords0 # return vcoords to normal
-                count += 1
-                if count == len(vcoords0 - mask_design.nholes):
-                    print("Couldn't find a non-redundant design :(")
-                    return mask_design
-            if rcheck == 0: # if the new hole is non-redundant:
-                mask_design.xy_coords_cm = np.append(mask_design.xy_coords_cm, np.reshape(new_hole, [1,2]), axis=0) # add new hole to design
-                mask_design.nholes += 1 # increase nholes
-                mask_design.xy_coords_m = mask_design.xy_coords_cm / 100
-                break # break out of while loop, go on to next hole
-    #print("Yay! Mask design is non-redundant. Plotting design...")
-    #plot_design(mask_design, aperture)
-    return mask_design # return updated design
-
-########################################################################################################
-########################################################################################################
 ########################################################################################################
 
 def save_design(mask_design):
@@ -298,77 +132,6 @@ def save_design(mask_design):
 ########################################################################################################
 ########################################################################################################
 ########################################################################################################
-
-#def calc_uv_param(mask_design):
-    #n = 100
-    #coordsx, coordsy = np.meshgrid(np.linspace(0, 1000, int(np.sqrt(n))), np.linspace(0, 500, int(np.sqrt(n))))
-    #coordsx2, coordsy2 = np.meshgrid(np.linspace(500, 1000, int(np.sqrt(n)/2)), np.linspace(0, 1000, int(np.sqrt(n)/2)))
-    #coords = np.vstack([coordsx.ravel(), coordsy.ravel()]).T
-    #coords2 = np.vstack([coordsx2.ravel(), coordsy2.ravel()]).T
-    #xy = np.loadtxt('/Users/kenzie/Desktop/Laniakea/Finalized_mask_pipeline/masks/SPIE_2024/9hole_nirc2_xycoords_cent.txt')
-    #coords = np.append(mask_design.uv_coords, -mask_design.uv_coords, axis=0)
-
-    #plt.figure()
-    #plt.scatter(coords[:, 0], coords[:, 1])
-    #plt.show()
-
-    #flatx = np.sort(coords[:, 0])
-    #flaty = np.sort(coords[:, 1])
-
-    #p, resid, n1, n2, n3  = np.polyfit(flatx, flaty, deg=1, full=True)
-    #print(p[1])
-    #print(resid) # / (len(flatx) - 2))
-
-    #plt.figure()
-    #plt.scatter(np.arange(-10, 10), np.arange(-10, 10))
-    #plt.scatter(flatx, flaty)
-    #plt.show()
-
-    return(p[0], p[1], resid)
-
-def plot_ps(mask_design):
-    coords = np.append(mask_design.uv_coords, -mask_design.uv_coords, axis=0)
-    #flatx = np.sort(coords[:, 0])
-    #flaty = np.sort(coords[:, 1])
-    #plt.figure()
-    #plt.scatter(np.arange(-10, 10), np.arange(-10, 10))
-    #plt.scatter(flatx, flaty)
-    #plt.show()
-
-    #xlocs = np.linspace(-10, 10, num_bins)
-    #ylocs = np.linspace(-10, 10, num_bins)
-
-    #plt.figure()
-    #plt.scatter(coords[:, 0], coords[:, 1])
-    #for i in range(len(xlocs)):
-    #    plt.axvline(xlocs[i], -10, 10)
-    #    plt.axhline(ylocs[i], -10, 10)
-    #plt.show()
-
-    dim = 200
-
-    stuff = np.zeros([dim, dim])
-    x = np.round(np.linspace(-10, 10, dim+1), 1)
-    coords2 = np.round(coords, 1)
-
-    for i in range(dim):
-        for j in range(dim):
-            for k in coords2:
-                if ([x[i], x[j]] == k).all() == 1:
-                    stuff[i, j] = 1
-
-    #plt.figure()
-    #plt.imshow(stuff)
-    #plt.show()
-
-    #params, covar, fit, stdx, stdy = fit_ps_fft(stuff)
-    #dim = int(np.sqrt(len(fit)))
-    #print(np.abs(stdx - stdy))
-    #plt.figure()
-    #plt.imshow(fit.reshape(dim, dim))   
-    #plt.colorbar()
-    #plt.show()
-    return stuff
 
 def fit_ps_fft(stuff):
     fft = np.fft.fftshift(np.fft.fft2(np.fft.fftshift(np.pad(stuff, 500))))
@@ -442,28 +205,3 @@ def fit_ps_fft(stuff):
 ########################################################################################################
 ########################################################################################################
     
-def calc_spacing(mask_design):
-
-    num_bins = 15
-    uvs = np.append(mask_design.uv_coords, -mask_design.uv_coords, axis=0)
-
-    xpos0 = -10
-    ypos0 = -10
-    count = 0
-    my_bins = np.zeros(num_bins)
-    bin_size = int(20 / int(np.sqrt(num_bins)))
-
-    for i in range(int(np.sqrt(num_bins))):
-        xpos = -10 + (i + 1) * bin_size
-        ypos0 = -10
-        for j in range(int(np.sqrt(num_bins))):
-            ypos = -10 + (j + 1) * bin_size
-            for k in uvs:
-                #print('is ' + str(k) + ' within ' + str(xpos0) + ' <= x < ' + str(xpos) + ' and ' + str(ypos0) + ' <= y < ' + str(ypos) + '?')
-                if k[0] < xpos and k[0] >= xpos0 and k[1] < ypos and k[1] >= ypos0:
-                    my_bins[count] += 1
-            count += 1
-            ypos0 = ypos
-        xpos0 = xpos
-
-    return(np.var(my_bins), num_bins)
